Moved channel switcher to seperate file
